Move homepage modules to separate components
